services/evaluator/requirements.txt
python-dotenv

git+https://github.com/Manga-ai-project/toolbox-logging.git